requirements.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.30
finance-datareader>=0.9.50

# Technical Analysis
ta>=0.10.2

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
tensorflow>=2.14.0

# Phase D: AI Chatbot (Gemini API)
google-generativeai>=0.8.0

# Visualization
matplotlib>=3.7.0
plotly>=5.17.0
mplfinance

# Web Dashboard
streamlit>=1.28.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0

# Phase 2 - News & Sentiment Analysis
beautifulsoup4>=4.12.0
feedparser>=6.0.10
textblob>=0.17.0
vaderSentiment>=3.3.2
transformers>=4.30.0
torch

# Phase 4 - Real-time Data
websockets>=12.0
